Added in shoreline
Shoreline added as a straight line for now at y =-1500. The plots have been adjusted to show the shoreline. In reality the shoreline would be a lot further out than 1500m but the plots wouldn't be as visually pleasing so it may be worth introducing some scaling factors instead when introducing the export cables.