Unconstrained substation from [0,0]
Substation now exists as an optimizable parameter within the optimization function.
Also added constraints so that the substation and turbines can't be within a certain distance from the shoreline
Fixed a problem with the export cable cost not being transferred over correctly to the post optimized layout. It was using the same value pre-optimization instead of the value from the optimized layout.
Also fixed the calculation for the minimum distance. It used pluses where it should be minuses when finding x and y