Removed scaling
Removed the scaling to make plots look nice. Instead are using the actual figures and are creating two separate plots. One plot showing the wind farm and another plot below it showing the export cable connection to the shoreline.
Also temporarily reverted back to a straight line for the shore line