General bug fixes
Fixed an issue with the incorrect plotting of the post optimize substation connection. The plot was incorporating the original substation position instead of the optimized position due to the substation not being an input into the used plotting function.
Fixed an error with the calculation of the inter array cabling costs. This was occurring simply because a part of the code had accidentally been deleted previously. This has since been reincorporated into the code.
Fixed an issue with the optimization finishing too quickly which caused less than optimal layouts to be produced. This was done by increasing 'eps' in the minimize function from 50 to 100.
Updated the initial distance from shore condition to consider the closest turbine instead of the farms centre.
Changed the colour bar limits back to considering the min and max limits as the min and max depths of the considered turbine positions.
Updated the bnds to relate to the limits of the bathymetry data.