Displayed 1st plot on GUI
Initial turbine position plot is now displayed on a page after the model parameters page.
This is currently in progress as it only produces the initial plot based off of the default user defined values. If you were to change these values on the user defined values page nothing will happen as of yet.